WIP: Move models out of core, and integrate prosit1 and Djura msms model there along with dummy
