+++ CITATION.cff
@@ -37,9 +37,9 @@ authors:
     orcid: "https://orcid.org/0000-0003-4096-0260"
   -
 
-date-released: 2024-06-07
+date-released: 2024-12-31
 doi: 10.5281/zenodo.7890716
-version: "0.4.0"
+version: "0.5.0"
 repository-code: "https://github.com/https://github.com/aiproteomics/aiproteomics"
 keywords:
   - machine learning
