+++ aiproteomics/__init__.py
@@ -1,1 +1,1 @@
-__VERSION__ = "0.4.0"
+__VERSION__ = "0.5.0"
