--- aiproteomics/frag/models/.gitignore
@@ -1,1 +1,0 @@
-/tutorial
