--- aiproteomics/frag/models/__init__.py
@@ -1,1 +1,0 @@
-from .early_fusion_transformer import *
