--- aiproteomics/modelgen/prosit1/__init__.py
@@ -1,4 +1,0 @@
-from . import constants
-from . import layers
-from . import losses
-from . import sanitize
