--- aiproteomics/rt/models/__init__.py
@@ -1,1 +1,0 @@
-from .transformer_rt import *
